custom protocol detection working

+++ core/src/main/java/me/mrnavastar/protoweaver/ProtoConstants.java
@@ -3,7 +3,7 @@ package me.mrnavastar.protoweaver;
 public class ProtoConstants {
 
     public static final int PROTOWEAVER_ID = 0x0000;
-    public static final int PROTOWEAVER_MAGIC_BYTE = 99;
+    public static final byte PROTOWEAVER_MAGIC_BYTE = 99;
 
     //Built in protocol IDs
     public static final int PROTO_MESSAGE_ID = 0x0001;
